backend/pyproject.toml
[tool.poetry]
name = "qrcode_generator"
version = "0.1.0"
description = "A simple qrcode generator"
authors = ["otonielnn <otoniel_2002@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.112.2"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pillow = "^10.1.0"
qrcode = {extras = ["pil"], version = "^7.4.2"}
pydantic = "^2.5.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.3"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.21.0"
pytest-xdist = "^3.0.0"
httpx = "^0.24.0"
psutil = "^5.9.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"
pre-commit = "^3.0.0"
taskipy = "^1.13.0"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
run = 'fastapi dev main.py'

pre_test = 'task lint'
test = 'pytest --cov=. --cov-report=term-missing --cov-report=html -vv'
post_test = 'coverage html'

lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"